+++ upcoming-release-notes/6150.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [MatissJanis]
+---
+
+Extract schedules form into a separate component
